Improvements addressing various pull-request comments.
