+++ jgrapht-core/src/test/java/org/jgrapht/alg/color/RandomGreedyColoring2Test.java
@@ -1,0 +1,49 @@
+/*
+ * (C) Copyright 2017-2017, by Dimitrios Michail and Contributors.
+ *
+ * JGraphT : a free Java graph-theory library
+ *
+ * This program and the accompanying materials are dual-licensed under
+ * either
+ *
+ * (a) the terms of the GNU Lesser General Public License version 2.1
+ * as published by the Free Software Foundation, or (at your option) any
+ * later version.
+ *
+ * or (per the licensee's choosing)
+ *
+ * (b) the terms of the Eclipse Public License v1.0 as published by
+ * the Eclipse Foundation.
+ */
+package org.jgrapht.alg.color;
+
+import java.util.Random;
+
+import org.jgrapht.Graph;
+import org.jgrapht.alg.interfaces.VertexColoringAlgorithm;
+import org.jgrapht.graph.DefaultEdge;
+
+/**
+ * Coloring tests
+ * 
+ * @author Dimitrios Michail
+ */
+public class RandomGreedyColoring2Test
+    extends BaseColoringTest
+{
+
+    final long seed = 15;
+
+    @Override
+    protected VertexColoringAlgorithm<Integer> getAlgorithm(Graph<Integer, DefaultEdge> graph)
+    {
+        return new RandomGreedyColoring<>(graph, new Random(seed));
+    }
+
+    @Override
+    protected int getExpectedResultOnDSaturNonOptimalGraph()
+    {
+        return 4;
+    }
+
+}
