+++ jgrapht-core/src/test/java/org/jgrapht/alg/color/SaturationDegreeColoringTest.java
@@ -1,0 +1,76 @@
+/*
+ * (C) Copyright 2017-2017, by Dimitrios Michail and Contributors.
+ *
+ * JGraphT : a free Java graph-theory library
+ *
+ * This program and the accompanying materials are dual-licensed under
+ * either
+ *
+ * (a) the terms of the GNU Lesser General Public License version 2.1
+ * as published by the Free Software Foundation, or (at your option) any
+ * later version.
+ *
+ * or (per the licensee's choosing)
+ *
+ * (b) the terms of the Eclipse Public License v1.0 as published by
+ * the Eclipse Foundation.
+ */
+package org.jgrapht.alg.color;
+
+import static org.junit.Assert.assertEquals;
+
+import java.util.Arrays;
+import java.util.Map;
+
+import org.jgrapht.Graph;
+import org.jgrapht.Graphs;
+import org.jgrapht.alg.interfaces.VertexColoringAlgorithm;
+import org.jgrapht.alg.interfaces.VertexColoringAlgorithm.Coloring;
+import org.jgrapht.graph.DefaultEdge;
+import org.jgrapht.graph.Pseudograph;
+import org.junit.Test;
+
+/**
+ * Coloring tests
+ * 
+ * @author Dimitrios Michail
+ */
+public class SaturationDegreeColoringTest
+    extends BaseColoringTest
+{
+
+    @Override
+    protected VertexColoringAlgorithm<Integer> getAlgorithm(Graph<Integer, DefaultEdge> graph)
+    {
+        return new SaturationDegreeColoring<>(graph);
+    }
+
+    @Override
+    protected int getExpectedResultOnDSaturNonOptimalGraph()
+    {
+        return 4;
+    }
+
+    @Test
+    public void testSaturationDegree()
+    {
+        Graph<Integer, DefaultEdge> g = new Pseudograph<>(DefaultEdge.class);
+        Graphs.addAllVertices(g, Arrays.asList(1, 2, 3, 4, 5, 6));
+        g.addEdge(2, 3);
+        g.addEdge(4, 5);
+        g.addEdge(4, 6);
+        g.addEdge(5, 6);
+        g.addEdge(5, 3);
+
+        Coloring<Integer> coloring = new SaturationDegreeColoring<>(g).getColoring();
+        assertEquals(3, coloring.getNumberColors());
+        Map<Integer, Integer> colors = coloring.getColors();
+        assertEquals(0, colors.get(1).intValue());
+        assertEquals(0, colors.get(2).intValue());
+        assertEquals(1, colors.get(3).intValue());
+        assertEquals(1, colors.get(4).intValue());
+        assertEquals(0, colors.get(5).intValue());
+        assertEquals(2, colors.get(6).intValue());
+    }
+
+}
